+++ src/EvolvingGraphs.jl
@@ -18,11 +18,11 @@ AbstractEvolvingGraph, key, make_node, index, rev, AttributeDict,
 TimeGraph, EvolvingGraph, WeightedEvolvingGraph, AttributeEvolvingGraph,
 
 # graph functions
-add_node!, add_edge!, add_graph!,
+add_node!, add_edge!, add_graph!, rm_edge!, has_edge, has_node,
 nodes, num_nodes, edges, num_edges,
 source, target, matrices, num_matrices, 
 timestamps, num_timestamps, reduce_timestamps!,
-has_node, out_neighbors, is_directed,
+out_neighbors, is_directed,
 time_graph, evolving_graph, weighted_evolving_graph,
 attribute_evolving_graph, attributesvec, attributes,
 matrix, spmatrix, 
@@ -37,19 +37,20 @@ katz_centrality, random_time_graph, random_evolving_graph,
 Path, TemporalPath, shortest_temporal_path, shortest_temporal_distance,
 shortest_path, shortest_distance,
 
-# util
+# sort slice
 issorted, sorttime!, sorttime, slice!, slice,
 
 # example
 build_evolving_graph
 
 include("common.jl")
+include("util.jl")
  
 include("graphs/time_graph.jl") 
 include("graphs/evolving_graph.jl")
 include("graphs/weighted_evolving_graph.jl")
 include("graphs/attribute_evolving_graph.jl")
-include("util.jl")
+include("sort_slice.jl")
 
 include("io.jl")
 include("show.jl")